refactor(visual-regions): split caption attachment helpers
Separate caption text recognition, caption extraction, and candidate scoring from the attachment facade so visual-region caption heuristics can evolve in smaller modules.
